app/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
dj-rest-auth==2.2.4
django-cors-headers==3.11.0
django-environ==0.9.0
django-filter==21.1
cryptography==37.0.4


asgiref==3.5.2
backports-zoneinfo==0.2.1 ; python_version < "3.9"
cffi==1.15.1





psycopg2-binary==2.9.3
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
